+++ nmanga/cmd.py
@@ -37,7 +37,7 @@ from .cli.manual_split import manual_split
 from .cli.merge_chapters import merge_chapters
 from .cli.releases import prepare_releases, prepare_releases_chapter
 from .cli.spreads_manager import spreads
-from .constants import __author__, __name__, __version__
+from ._metadata import __author__, __name__, __version__
 from .term import get_console
 
 console = get_console()