+++ nmanga/constants.py
@@ -22,11 +22,38 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__name__ = "nmanga"
-__author__ = "noaione"
-__author_email__ = "noaione@n4o.xyz"
-__version__ = "0.9.1"
-__license__ = "MIT"
-__copyright__ = "Copyright (c) 2022-present noaione"
+from dataclasses import dataclass
 
-__all__ = ("__name__", "__author__", "__version__", "__license__", "__copyright__")
+__all__ = (
+    "MangaPublication",
+    "MANGA_PUBLICATION_TYPES",
+    "TARGET_FORMAT",
+    "TARGET_FORMAT_ALT",
+    "TARGET_TITLE",
+)
+
+
+@dataclass
+class MangaPublication:
+    image: str
+    """Used in the image filename"""
+    archive: str
+    """Used in the archive filename"""
+    description: str
+    """Description, this will be used in the config command for what it means."""
+
+
+MANGA_PUBLICATION_TYPES = {
+    "digital": MangaPublication("dig", "Digital", "Digital manga release from Amazon, etc."),
+    "magazine": MangaPublication("mag", "c2c", "Magazine release, can be used for scanlation batch."),
+    "scan": MangaPublication("c2c", "c2c", "Scanned manga release, physical turned to digital."),
+    "web": MangaPublication(
+        "web", "Digital", "Webcomic release, from many kind of website that are not using volume format."
+    ),
+    "mix": MangaPublication("mix", "Digital", "Mixed release format"),
+    "none": MangaPublication("", "Digital", "No publication type used in image filename, Digital in archive filename"),
+}
+
+TARGET_FORMAT = "{mt} - c{ch}{chex} ({vol}) - p{pg}{ex}{pt} [{t}] [{pb}] [{c}]"  # noqa
+TARGET_FORMAT_ALT = "{mt} - c{ch}{chex} ({vol}) - p{pg}{ex}{pt} [{pb}] [{c}]"  # noqa
+TARGET_TITLE = "{mt}{vol} ({year}) ({pt}) {cpa}{c}{cpb}"